requirements.txt
Flask==3.0.3
flask_sqlalchemy==3.1.1
pandas==2.2.2
pytest==8.3.2
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.32
